requirements.txt
altgraph==0.17
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
google-api-core==1.31.0
google-api-python-client==2.13.0
google-auth==1.33.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
macholib==1.14
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller==4.5
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
simplejson==3.17.3
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.6
tqdm==4.61.2
